==> code/02_random_forest.qmd <==
---
title: "Data Science Applied to Ag - Final Project - ML"
format:
  html:
    embed-resources: true
    toc: true
    theme: cerulean
author: Md Shakir Moazzem, Umar Munir
---
# Introduction  

This script contains ML Workflow with random forest. 

# Setup  
## Install and Load relevant packages  

The following code chunk will load necessary packages.

```{r Setup, message=F, warning=F}

# Installing packages

#install.packages("tidyverse")
#install.packages("readxl")
#install.packages("janitor")
#install.packages("dplyr")
#install.packages("tidyr")
#install.packages("readr")
#install.packages("lubridata")
#install.packages("stringr")
#install.packages("ranger") 
#install.packages("vip")
#install.packages("ranger")
#install.packages("finetune")

# Loading packages 

library(readxl) 
library(janitor) 
library(dplyr) 
library(tidyr) 
library(readr) 
library(lubridate)
library(stringr)
library(tidymodels)
library(tidyverse) 
library(vip)
library(ranger)
library(finetune)

```
 
 
 ## Data import

The following code chunk will import the finali wrangled  data file i.e., "weather_monthsum.csv".

```{r import, message=F, warning=F}
# Going to do some wrangling on the go with the file
weather_rf <- read_csv("../data/weather_monthsum.csv") %>%
  rename(yield = adjusted_yield) 

weather_rf

#view(weather_rf)


```
 
 
# Machine Learning workflow  

## 1. Pre-processing  

The following code chunk will split off 2024 for final prediction. 

```{r}
# Splitting data between known (2014-2023) and unknown (2024) containing  NA as yield values for future prediction
weather_to_predict <- weather_rf %>% filter(is.na(yield)) #prediction

weather_complete     <- weather_rf %>% filter(!is.na(yield))
```

## a. Data split  

The following code chunk will conduct train/test split for the data of the years 2014–2023.

```{r}
# Initial data split (70/30) on the known data (2014-2023)
set.seed(931735)
weather_split_rf <- initial_split(
  weather_complete,
  prop   = 0.7,
  strata = yield
)

```



The following code chunk will conduct setting train set.

```{r weather_train, message=F, warning=F}
# Setting train set 
weather_train_rf <- training(weather_split_rf)

weather_train_rf
```

The following code chunk will conduct setting test split.


```{r weather_test, message=F, warning=F}
# Setting test split
weather_test_rf <- testing(weather_split_rf)

weather_test_rf
```

 
 
### b. Distribution of target variable "yield"

The following code chunk will create a density plot to compare target variable "yield" in the training and test set.

```{r distribution, message=F, warning=F}
ggplot() +
  geom_density(data = weather_train_rf, 
               aes(x = yield),
               color = "red") +
  geom_density(data = weather_test_rf, 
               aes(x = yield),
               color = "blue") 
  
```


### c. Data processing with recipe

The following code chunk will conduct data processing with recipe.

```{r weather_recipe}
weather_recipe_rf <-
  # Defining predicted and predictor variables
  recipe(yield ~ .,
         data = weather_train_rf) %>%
  # Removing year, site, hybrid, and non-growing season  
    step_rm(year, 
            site, 
            hybrid,
            matches("Jan|Feb|Mar|Nov|Dec")) 
  
weather_recipe_rf
```


The following code chunk will prep the recipe to estimate any required statistics.

```{r weather_prep}

weather_prep_rf <- weather_recipe_rf %>%
  prep()

weather_prep_rf
```

 
## 2. Training

### a. Model specification

The following code chunk will conduct model specification.
  
```{r rf_spec}

rf_spec <- 
  # Specifying random forest as our model type, asking to tune the hyperparameters
  rand_forest(trees = tune(),
              mtry = tune() 
              ) %>%
    # Specify the engine (= package)
    set_engine("ranger") %>% #specifying "ranger" as the engine/package to run random forest 
    # Specifying mode  
    set_mode("regression") #random forest can handle both regression (when y is numerical) and classification (when y is categorical) #Here, we are specifying "set_mode("regression")" because our y variable is numerical [continuous]

rf_spec

```

# Bundle recipe & model into workflow

The following code chunk will bundle recipe & model into workflow.

```{r}
# Bundling into a workflow
library(workflows)
rf_wf <- workflow() %>%
  add_recipe(weather_recipe_rf) %>% # adding preprocessing
  add_model(rf_spec) # adding XGBoost spec
```

# Hyperparameter tuning

The following code chunk will conduct 3-fold cross-validation to evaluate model performance during tuning.

```{r resampling_foldcv}
set.seed(34549)
resamples_rf <- vfold_cv(weather_train_rf, 
                              v = 3) #becuase of limited time and computing power

resamples_rf
resamples_rf$splits[[1]]
resamples_rf$splits[[2]]
```

# Hyperparameter grid search with simulated annealing

The following code chunk will conduct hyperparameter grid search with simulated annealing

```{r simulated annealing}
# Tuning with simulated annealing (3 because of limited time and computational power)
library(finetune)
set.seed(76544)
rf_tune_res <- tune_sim_anneal(
  rf_wf,
  resamples = resamples_rf,
  iter = 3,
  control = control_sim_anneal(save_workflow = FALSE)
)


beepr::beep() 
```

# Selecting best hyperparameters

The following code chunk will select best hyperparameters.

```{r rmse based best model}
# Selecting the best hyperparameters by RMSE
best_params <- rf_tune_res %>% 
  select_by_pct_loss("trees",
                     metric = "rmse",
                     limit = 2)

```

# Finalizing workflow & fitting full training set

The following code chunk will finalize workflow & fit full training set.

```{r workflow finalizaiton}
# Finalizing the workflow
final_rf_wf <- finalize_workflow(rf_wf, best_params)

# Fitting on all of the training data (70% of the data split)
final_rf_fit <- fit(final_rf_wf, data = weather_train_rf)
```

# Evaluation on held-out test set 

The following code chunk will conduct evaluation on the held-out test set.

```{r}
# Evaluating on the held‐out test set (30% of the data split)
library(yardstick)
test_preds <- predict(final_rf_fit, weather_test_rf) %>%
  bind_cols(weather_test_rf)
test_preds %>%
  rmse(truth = yield, estimate = .pred) %>%
  bind_rows(
    rsq(test_preds, truth = yield, estimate = .pred)
  )
```

# Prediction

The following code chunk will conduct predictioin for the 2024 rows.

```{r}

# Now I am going to make predictioin for the 2024 rows
pred_2024 <- predict(final_rf_fit, weather_to_predict) %>%
  bind_cols(weather_to_predict)


# Quick Wrangling
pred_2024_w <- pred_2024 %>% 
  dplyr::select(year, site, hybrid, .pred) %>% 
  rename(pred_yield_mg_ha = .pred) %>% 
  distinct()

#Final data
pred_2024_w
#write_csv(pred_2024_w, "predicted_2024_yields.csv")

```
 
 
# Predicted vs Observed plot

The following code chunk will create Predicted vs Observed plot.

```{r observed vs predicted}
# Plotting Predicted vs Observed Yield for the test set
ggplot(test_preds, aes(x = yield, y = .pred)) +
  geom_point(color = "blue", alpha = 0.6) +
  geom_abline(slope = 1, intercept = 0, color = "red", linetype = "dashed") +
  labs(
    title = "Predicted vs Observed Yield",
    x = "Observed Yield (Mg/ha)",
    y = "Predicted Yield (Mg/ha)"
  ) +
  theme_minimal()

```

# Variable importance plot

The following code chunk will create variable importance plot. 

```{r VIP plot}
# Variable Importance Plot from the final fitted model
library(vip)

# Extracting the fitted ranger model from final_rf_fit
rf_model <- extract_fit_parsnip(final_rf_fit)$fit

# Generating variable importance plot
vip(rf_model, num_features = 20) +
  labs(title = "Variable Importance from Random Forest")

```

 
# Combining with submission

The following code chunk will sort the columns in our predicted 2024 yields inside "pred_2024_final" based on the "testing_submission".

```{r}

#pred_2024_final <- read.csv("../output/rf_predicted_2024_yields.csv") 
test_submission <- read.csv("../data/testing/testing_submission.csv")

pred_2024_filtered <- pred_2024_w %>%
  distinct(year, site, hybrid, .keep_all = TRUE) %>%
  semi_join(test_submission, by = c("year", "site", "hybrid"))

pred_2024_filtered

#write.csv(pred_2024_filtered, "randomforest_yieldprediction_2024.csv") 
```
